+++ src/canonmn_int.cpp
@@ -2730,7 +2730,7 @@ namespace Exiv2 {
         ExifData::const_iterator pos = metadata->findKey(key);
         if (pos != metadata->end() && pos->value().count() >= 3 && pos->value().typeId() == unsignedShort) {
             float fu = pos->value().toFloat(2);
-            if (fu != 0.0) {
+            if (fu != 0.0f) {
                 float fl = value.toFloat(1) / fu;
                 std::ostringstream oss;
                 oss.copyfmt(os);
@@ -2845,8 +2845,7 @@ namespace Exiv2 {
     }
 
     //! convertFocalLength to a human readable string
-    void convertFocalLength(LensTypeAndFocalLengthAndMaxAperture& ltfl,
-                            double divisor)
+    void convertFocalLength(LensTypeAndFocalLengthAndMaxAperture& ltfl, float divisor)
     {
         std::ostringstream oss;
         oss << std::fixed << std::setprecision(0);
@@ -2870,8 +2869,9 @@ namespace Exiv2 {
         ltfl.lensType_ = value.toLong();
 
         extractLensFocalLength(ltfl, metadata);
-        if (ltfl.focalLengthMax_ == 0.0) return os << value;
-        convertFocalLength(ltfl, 1.0);
+        if (ltfl.focalLengthMax_ == 0.0f)
+            return os << value;
+        convertFocalLength(ltfl, 1.0f);
 
         ExifKey key("Exif.CanonCs.MaxAperture");
         ExifData::const_iterator pos = metadata->findKey(key);
@@ -2907,8 +2907,9 @@ namespace Exiv2 {
         ltfl.lensType_ = value.toLong();
 
         extractLensFocalLength(ltfl, metadata);
-        if (ltfl.focalLengthMax_ == 0.0) return os << value;
-        convertFocalLength(ltfl, 1.0);
+        if (ltfl.focalLengthMax_ == 0.0f)
+            return os << value;
+        convertFocalLength(ltfl, 1.0f);
 
         if (ltfl.focalLength_.empty()) return os << value;
 
@@ -2929,14 +2930,15 @@ namespace Exiv2 {
         ltfl.lensType_ = value.toLong();
 
         extractLensFocalLength(ltfl, metadata);
-        if (ltfl.focalLengthMax_ == 0.0) return os << value;
-        convertFocalLength(ltfl, 1.0); // just lens
+        if (ltfl.focalLengthMax_ == 0.0f)
+            return os << value;
+        convertFocalLength(ltfl, 1.0f);  // just lens
         const TagDetails* td = find(canonCsLensType, ltfl);
         if (!td) {
-            convertFocalLength(ltfl, 1.4); // lens + 1.4x TC
+            convertFocalLength(ltfl, 1.4f); // lens + 1.4x TC
             td = find(canonCsLensType, ltfl);
             if (!td) {
-                convertFocalLength(ltfl, 2.0); // lens + 2x TC
+                convertFocalLength(ltfl, 2.0f); // lens + 2x TC
                 td = find(canonCsLensType, ltfl);
                 if (!td) return os << value;
             }
@@ -2983,7 +2985,8 @@ namespace Exiv2 {
         }
 
         float fu = value.toFloat(2);
-        if (fu == 0.0) return os << value;
+        if (fu == 0.0f)
+            return os << value;
         float len1 = value.toLong(0) / fu;
         float len2 = value.toLong(1) / fu;
         std::ostringstream oss;
@@ -3006,7 +3009,7 @@ namespace Exiv2 {
         std::ios::fmtflags f( os.flags() );
         if (   value.typeId() == unsignedShort
             && value.count() > 0) {
-            os << exp(canonEv(value.toLong()) / 32 * log(2.0)) * 100.0;
+            os << std::exp(canonEv(value.toLong()) / 32 * std::log(2.0f)) * 100.0f;
         }
         os.flags(f);
         return os;
@@ -3020,7 +3023,7 @@ namespace Exiv2 {
         if (   value.typeId() == unsignedShort
             && value.count() > 0) {
             // Ported from Exiftool by Will Stokes
-            os << exp(canonEv(value.toLong()) * log(2.0)) * 100.0 / 32.0;
+            os << std::exp(canonEv(value.toLong()) * std::log(2.0f)) * 100.0f / 32.0f;
         }
         os.flags(f);
         return os;
